fixed PR review comments, moved useGetNewChatId into useMakeChatGPTAPIInstance
