+++ src/redux/resultsSlice.ts
