+++ src/redux/settingsSlice.ts
